dont read from S3 on import
